+++ static/__tests__/dtale/reshape/DataViewer-reshape-transpose-test.jsx
@@ -113,6 +113,13 @@ describe("DataViewer tests", () => {
           .instance()
           .onChange([{ value: "col2" }]);
         result
+          .find("div.modal-body")
+          .find("div.row")
+          .last()
+          .find("button")
+          .last()
+          .simulate("click");
+        result
           .find("div.modal-footer")
           .first()
           .find("button")
@@ -122,10 +129,11 @@ describe("DataViewer tests", () => {
           result.update();
           t.equal(result.find(Reshape).length, 1, "should hide reshape");
           result
-            .find(Reshape)
             .find("div.modal-body")
-            .find("button")
+            .find("div.row")
             .last()
+            .find("button")
+            .first()
             .simulate("click");
           result
             .find("div.modal-footer")